Use LanguageConverter in SearchableNamespaceListBuilder
Warning, I'm not 100% sure if this is the best approach. I looked
at a few examples and tried to reduce the code to the absolute
minimum. This especially means there is no ->hasVariants() or
->hasVariant( $variant ) any more. Instead the LanguageConverter
is always used, even in languages where it is not needed. Is this
a problem, e.g. because of performance concerns?

I'm also confused why …&uselang=zh still shows the namespace names
in English. Is this intentional?

Bug: T401499
Change-Id: Ia34d49b20215f219ca157948211f9cdf4b15f7ca