Consistent use of spaces and : markup in long messages
This extension has several long "Description" messages,
which are shown in the search form when the user
clicks the "information" icons next to the fields.

Most of those messages have consistent markup
with the ";" and ":" characters that create definition lists.
There is usually no ":" after "Description", and there is
usually no space after the ":" character.
A few messages were different. This patch makes
them all consistent.

Change-Id: I362af5e503bc94594017b5fa569a27a71c46e4cc