Fix broken calls to the internal query cache
These calls appear to be broken ever since this was written for the
first time in 2018 via I17a3016. The effect was that some of this was
never correctly cached but potentially queried over and over again.
Luckily this is rare and can only happen in two situations:

1. When a title is invalid, e.g. because of invalid characters. This
never made it to the cache.

2. When the network is extremely slow it was possible to have multiple
API requests running because the intermediate "pending" state was never
correctly cached. This wastes the network but doesn't have a visible
effect for the user.

More cleanups for this confusing code will follow in If8f829f.

Change-Id: Ib547d0fc1f0111240dce0ef09d605b2f9cec4141