Replace strings with bools in internal query cache
We need three states: If a page exists or not, and a third state for
titles that are currently queried. While strings work, they have
disadvantages. For example, what do "ok", "no" and so on even mean?

I think a boolean "exists()" method makes all this quite a bit more
readable.

Change-Id: Ic856177d9f4e520e409c7b8a00313195a749e9aa